+++ src/main/java/flab/delideli/dao/PaymentDao.java
@@ -16,4 +16,6 @@ public interface PaymentDao {
 	void updatePaymentStatusUnableCancel(long orderId);
 
 	void updatePaymentStatusCanceledByOwner(long orderId);
+
+	long selectPaymentId(long orderId);
 }
