+++ src/main/java/flab/delideli/service/OrderOwnerService.java
@@ -1,12 +1,11 @@
 package flab.delideli.service;
 
-import flab.delideli.dao.OwnerDao;
+import flab.delideli.dao.OrderDao;
 import flab.delideli.dao.PaymentDao;
 import flab.delideli.dao.ShopDao;
 import flab.delideli.dto.OrderDTO;
 import flab.delideli.enums.OrderStatus;
 import lombok.RequiredArgsConstructor;
-import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class OrderOwnerService implements OwnerService {
 
-	private final OwnerDao ownerDao;
+	private final OrderDao orderDao;
 	private final ShopDao shopDao;
 	private final PaymentDao paymentDao;
 
@@ -24,7 +23,7 @@ public class OrderOwnerService implements OwnerService {
 		validateCorrectOwnerShop(userId, orderId);
 		OrderDTO orderDTO = getOrderDTO(orderId);
 		validateCorrectOrderStatus(orderDTO);
-		ownerDao.updateOrderStatusCooking(orderId);
+		orderDao.updateOrderStatusCooking(orderId);
 		paymentDao.updatePaymentStatusUnableCancel(orderId);
 	}
 
@@ -34,7 +33,7 @@ public class OrderOwnerService implements OwnerService {
 		validateCorrectOwnerShop(userId, orderId);
 		OrderDTO orderDTO = getOrderDTO(orderId);
 		validateCorrectOrderStatus(orderDTO);
-		ownerDao.updateOrderStatusCancel(orderId);
+		orderDao.updateOrderStatusCancel(orderId);
 		paymentDao.updatePaymentStatusCanceledByOwner(orderId);
 	}
 
@@ -45,12 +44,12 @@ public class OrderOwnerService implements OwnerService {
 		if (orderDTO.getOrderStatus() != OrderStatus.START_COOKING) {
 			throw new IllegalStateException("조리완료가 될 수 없는 주문입니다");
 		}
-		ownerDao.updateOrderStatusCookingComplete(orderId);
+		orderDao.updateOrderStatusCookingComplete(orderId);
 	}
 
 	/*주문이 들어온 곳이 사장님 가게의 주문이 맞는지 확인해야한다*/
 	private void validateCorrectOwnerShop(String userId, long orderId) {
-		long shopId = ownerDao.getShopIdInOrders(orderId);
+		long shopId = orderDao.getShopIdInOrders(orderId);
 		if (!userId.equals(shopDao.getOwnerIdInShops(shopId))) {
 			throw new IllegalArgumentException("잘못된 값이 들어왔습니다.");
 		}
@@ -65,7 +64,7 @@ public class OrderOwnerService implements OwnerService {
 	}
 
 	private OrderDTO getOrderDTO(long orderId) {
-		return ownerDao.selectOrderDTO(orderId);
+		return orderDao.selectOrderDTO(orderId);
 	}
 
 }
