--- src/main/resources/mappers/OwnerMapper.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="flab.delideli.dao.OwnerDao" >
-
-    <update id="updateOrderStatusCooking">
-        UPDATE orders
-        SET order_status = 'START_COOKING'
-        WHERE order_id = #{orderId}
-    </update>
-    
-    <select id="getShopIdInOrders" resultType="long">
-        SELECT shop_id
-        FROM orders
-        WHERE order_id = #{orderId}
-    </select>
-
-    <update id="updateOrderStatusCancel">
-        UPDATE orders
-        SET order_status = 'ORDER_CANCEL'
-        WHERE order_id = #{orderId}
-    </update>
-
-    <update id="updateOrderStatusCookingComplete">
-        UPDATE orders
-        SET order_status = 'COOKING_COMPLETE'
-        WHERE order_id = #{orderId}
-    </update>
-
-    <select id="selectOrderDTO" resultMap="selectOrderDTOs">
-        SELECT user_id, address, phone_number, pay_type, total_price, request, order_date, shop_id, order_status
-        FROM orders
-        WHERE order_id = #{orderId}
-    </select>
-
-    <resultMap id="selectOrderDTOs" type="flab.delideli.dto.OrderDTO">
-        <result column="user_id" property="userId"></result>
-        <result column="phone_number" property="phoneNumber"></result>
-        <result column="pay_type" property="paymentType"></result>
-        <result column="total_price" property="totalPrice"></result>
-        <result column="order_date" property="orderDate"></result>
-        <result column="shop_id" property="shopId"></result>
-        <result column="order_status" property="orderStatus"></result>
-    </resultMap>
-</mapper>
