+++ src/main/resources/mappers/PaymentMapper.xml
@@ -20,6 +20,12 @@
           AND user_id = #{userId}
     </select>
 
+    <select id="selectPaymentId" resultType="_long">
+        SELECT payment_id
+        FROM payments
+        WHERE order_id = #{orderId}
+    </select>
+
     <resultMap id="paymentDTO" type="flab.delideli.dto.PaymentDTO">
         <result column="order_id" property="orderId"></result>
         <result column="user_id" property="userId"></result>
